operatino/namespaced-require-server-transform: move interceptor from redirects, and add plugins catch
